+++ Integral.Games/Statistics/ConstantStatistic.cs
@@ -2,8 +2,8 @@
 {
     public sealed class ConstantStatistic : ReadOnlyStatistic
     {
-        public ConstantStatistic(int value = default) => Value = value;
+        public ConstantStatistic(float value = default) => Value = value;
 
-        public int Value { get; }
+        public float Value { get; }
     }
 }
